% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquasys.R
\name{+.aquasys}
\alias{+.aquasys}
\title{Modify an aquatic system by adding/replacing parameters and components}
\usage{
\method{+}{aquasys}(e1, e2)
}
\arguments{
\item{e1}{An object of class \code{aquasys}}

\item{e2}{A parameter or component to add to \code{e1}}
}
\description{
Modify an aquatic system by adding/replacing parameters and components
}
