% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{arithmetic}
\alias{arithmetic}
\title{Arithmetic}
\usage{
qty == qty, qty != qty, qty < qty, qty <= qty, qty > qty, qty >= qty

qty + qty, qty - qty

qty / qty

qty / number

qty * number

amount / volume = molarity, amount / molarity = volume, molarity * volume = amount

mass / MW = amount, mass / amount = MW, amount * MW = mass
}
\description{
These arithmetic operators for quantities in chemtools keep track of units in standard operations (i.e. they ALL take the SI prefix into consideration). Also note that all operations that result in a new quantity object automatically scale the new value using \code{\link{cht_best_metric}}.
}
\details{
\code{qty ==, !=, <, <=, >, >= qty} allows the comparison of quantities that are the same type (e.g. all mass).

\code{qty +- qty} allows the addition/subtraction of quantities that are the same type (e.g. all mass). Note that this operation also scales the new value using \code{\link{cht_best_metric}}.

\code{qty / qty} divide quantities of the same type. Returns plain numeric (i.e. the units are divided out).

\code{qty / number} divide quantity by a plain number. Returns the quantity automatically rescaled to the best metric.

\code{qty * number} multiply a quantity by a plain number (either way around is valid). Returns the quantity automatically rescaled to the best metric.

\code{amount / volume} divide an amount by a volume to create a molarity (concentration).

\code{amount / molarity} divide an amount by a molarity to create a volume.

\code{molarity * volume} multiply a molarity by a volume (or the other way around) to create an amount.

\code{mass / MW} divide a mass by a molecular weight to create an amount (mols).

\code{mass / amount} divice a mass by an amount (mols) to create a molecular weight.

\code{amount * MW} multiply an amount (mols) by a molecular weight to create a mass.
}
\examples{
cht_qty(1, "g") == cht_qty(1000, "mg") # TRUE
cht_qty(2, "mg") < cht_qty(5, "ng")  # FALSE
cht_qty(1000, "mg") + cht_qty(999, "g") # 1 kg
cht_qty(5, "mg") / cht_qty(1, "g") # 0.005
cht_qty(5, "mg") / 1e6 # 5 ng
cht_qty(5, "mg") * 1e6 # 5 kg
cht_qty(5, "nmol") / cht_qty(50, "mL") # 100 nM
cht_qty(5, "nmol") / cht_qty(100, "nM") # 50 mL
cht_qty(100, "nM") * cht_qty(50, "mL") # 5 nmol
cht_qty(10, "g") / qty (50, "g/mol") # 200 mM 
cht_qty(10, "g") / cht_qty(200, "mmol") # 50 g/mol
cht_qty(200, "mmol") * qty (50, "g/mol") # 10 g
}
