% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbonate_chem.R
\name{calculate_closed_system_pCO2}
\alias{calculate_closed_system_pCO2}
\title{CO2 pressure in a closed system}
\usage{
calculate_closed_system_pCO2(pH, TIC, V_gas, V_liquid,
  solubility = calculate_solubility("CO2", temperature),
  temperature = qty(25, "C"), pKa1 = 6.3, pKa2 = 10.3)
}
\arguments{
\item{pH}{the pH of the system}

\item{TIC}{total inorganic carbon in system (an amount quantity)}

\item{V_gas}{volume of the gas/headsapce (a volume quantity)}

\item{V_liquid}{volume of liquid (a volume quantity)}

\item{solubility}{the solubility of CO2, by default is determined based on the temperature of the system}

\item{temperature}{the temperature of the system, used to calculate the solubility constant}

\item{pKa1}{the acid dissociation constant for H2CO3*}

\item{pKa2}{the acid dissociation constant for bicarbonate (HCO3-)}
}
\value{
pCO2 (a pressure quantity)
}
\description{
Calculates the partial pressure of CO2 in the headspace of a closed system at equilibrium between gas and liquid.
}
