% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbonate_chem.R
\name{calculate_closed_system_TIC}
\alias{calculate_closed_system_TIC}
\title{Total inorganic carbon in a closed system}
\usage{
calculate_closed_system_TIC(pH, pCO2, V_gas, V_liquid,
  temperature = qty(25, "C"))
}
\arguments{
\item{pH}{the pH of the system}

\item{pCO2}{the partial pressure of CO2 (a pressure quantity)}

\item{V_gas}{volume of the gas/headsapce (a volume quantity)}

\item{V_liquid}{volume of liquid (a volume quantity)}

\item{temperature}{the temperature of the system, used to calculate the solubility constant}
}
\value{
total amount of inorganic carbon (an amount quantity)
}
\description{
Calculates total inorganic carbon carbon in a closed system at equlibrium with its headspace.
}
