% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbonate_chem.R
\name{calculate_closed_system_unbalanced_ions}
\alias{calculate_closed_system_unbalanced_ions}
\title{Calculate unbalanced ions of a closed system}
\usage{
calculate_closed_system_unbalanced_ions(pH, TIC, V_liquid, V_gas,
  solubility = calculate_solubility("CO2", temperature),
  temperature = qty(25, "C"), pKa1 = 6.3, pKa2 = 10.3, pKw = 14,
  buffer = qty(0, "M"), pKa)
}
\arguments{
\item{pH}{the pH of the system}

\item{TIC}{total inorganic carbon in system (an amount quantity)}

\item{V_liquid}{volume of liquid (a volume quantity)}

\item{V_gas}{volume of the gas/headsapce (a volume quantity)}

\item{solubility}{the solubility of CO2, by default is determined based on the temperature of the system}

\item{temperature}{the temperature of the system, used to calculate the solubility constant}

\item{pKa1}{the acid dissociation constant for H2CO3*}

\item{pKa2}{the acid dissociation constant for bicarbonate (HCO3-)}

\item{buffer}{[optional] total buffer concentration (a molarity quantity): a protonated buffer, make sure to add same amount to hard_ions IF a buffer salt is used (e.g. Na-buffer), and provide the appropriate pKa.}

\item{pKa}{buffer acid dissociation constant, required if \code{buffer} is provided}
}
\value{
unbalanced ion concentration (a molarity quantity), positive = excess cations, negative = excess anions
}
\description{
Function to calculate the unbalanced ions (e.g. Na+ added as NaOH or NaHCO3 minus Cl- added HCl) of a closed system that is carbonate buffered and has an optional additional buffer.
}
