% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbonate_chem.R
\name{calculate_DIC}
\alias{calculate_DIC}
\title{Function to calculate DIC}
\usage{
calculate_DIC(pH, pCO2, `H2CO3*` = solubility * pCO2,
  solubility = calculate_solubility("CO2", temperature),
  temperature = qty(25, "C"), pKa1 = 6.3, pKa2 = 10.3)
}
\arguments{
\item{pH}{the pH of the system}

\item{pCO2}{the partial pressure of CO2 (a pressure quantity)}

\item{H2CO3*}{the concentration of H2CO3* (a molarity quantity), by default is calculated automatically from pCO2}

\item{solubility}{the solubility of CO2, by default is determined based on the temperature of the system}

\item{temperature}{the temperature of the system, used to calculate the solubility constant}

\item{pKa1}{the acid dissociation constant for H2CO3*}

\item{pKa2}{the acid dissociation constant for bicarbonate (HCO3-)}
}
\value{
DIC concentration (a molarity quantity).
}
\description{
Calculates the concentration of dissolved inorganic carbon based on pH and H2CO3* (aqueous CO2 + H2CO3) concentration or CO2 pressure.
}
\examples{
cc_calculate_DIC(pH = 7.0, pCO2 = qty(0.4, "mbar"))
cc_calculate_DIC(pH = 7.0, pCO2 = qty(0.4, "mbar"), temperature = qty(0, "C"))
}
\concept{carbonate chemistry}
