% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_ideal_gas_amount}
\alias{calculate_ideal_gas_amount}
\title{Calculate the amount of an ideal gas}
\usage{
calculate_ideal_gas_amount(pressure, temperature, volume)
}
\arguments{
\item{pressure}{the pressure quantity of the gas}

\item{temperature}{the temperature quantity of the gas}

\item{volume}{the volume quantity of the gas}
}
\value{
amount quantity
}
\description{
Calculate the amount of an ideal gas at a specific pressure, temperature and volume
}
\seealso{
Other calculations: \code{\link{calculate_ideal_gas_molarity}},
  \code{\link{calculate_solubility}}
}
\concept{calculations}
