% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_ideal_gas_molarity}
\alias{calculate_ideal_gas_molarity}
\title{Calculate the concentration of an ideal gas}
\usage{
calculate_ideal_gas_molarity(pressure, temperature)
}
\arguments{
\item{pressure}{the pressure quantity of the gas}

\item{temperature}{the temperature quantity of the gas}
}
\value{
molarity quantity
}
\description{
Calculate the gas phase equivalent concentration of an ideal gas at a specific temperature (i.e. mol of gas / L of volume).
}
\seealso{
Other calculations: \code{\link{calculate_ideal_gas_amount}},
  \code{\link{calculate_solubility}}
}
\concept{calculations}
