% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbonate_chem.R
\name{calculate_open_system_pH}
\alias{calculate_open_system_pH}
\title{Calculate pH in open system (i.e. unlimited pCO2)}
\usage{
calculate_open_system_pH(pCO2, `H2CO3*` = solubility * pCO2,
  solubility = calculate_solubility("CO2", temperature),
  temperature = qty(25, "C"), pKa1 = 6.3, pKa2 = 10.3, pKw = 14,
  buffer = qty(0, "M"), pKa, unbalanced_ions = qty(0, "M"))
}
\arguments{
\item{pCO2}{the partial pressure of CO2 (a pressure quantity)}

\item{H2CO3*}{the concentration of H2CO3* (a molarity quantity), by default is calculated automatically from pCO2}

\item{solubility}{the solubility of CO2, by default is determined based on the temperature of the system}

\item{temperature}{the temperature of the system, used to calculate the solubility constant}

\item{pKa1}{the acid dissociation constant for H2CO3*}

\item{pKa2}{the acid dissociation constant for bicarbonate (HCO3-)}

\item{buffer}{[optional] total buffer concentration (a molarity quantity): a protonated buffer, make sure to add same amount to hard_ions IF a buffer salt is used (e.g. Na-buffer), and provide the appropriate pKa.}

\item{pKa}{buffer acid dissociation constant, required if \code{buffer} is provided}

\item{unbalanced_ions}{[optional] concentration of all unbalanced ions [units charge x a molarity quantity]. That means all charge multiplied cations from hard bases and included soft bases (carbonate and the specific buffer) e.g. mol/L Na that was added in the form of NaOH, NaHCO3 or as part of a Na-buffer salt. Also all charge multiplied (and -) anions from hard acids, e.g. -1 x mol/L Cl that was added as HCl or -2 x mol/L SO4 that was added as H2SO4.}
}
\value{
pH
}
\description{
Calculate pH in open system (i.e. unlimited pCO2)
}
