% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_solubility}
\alias{calculate_solubility}
\title{Calculate solubility of a gas}
\usage{
calculate_solubility(gas, temperature)
}
\arguments{
\item{gas}{of the gas}

\item{temperature}{temperature quantity}
}
\value{
solubility quantity
}
\description{
Calculate Henry's law solubility constant for a gas at a specific temperature. Henry's law constants are from: Sander, R. Compilation of Henry's law constants (version 4.0) for water as solvent. Atmos Chem Phys 15, 4399–4981 (2015). https://www.atmos-chem-phys.net/15/4399/2015/
}
\seealso{
Other calculations: \code{\link{calculate_ideal_gas_amount}},
  \code{\link{calculate_ideal_gas_molarity}}
}
\concept{calculations}
