% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities.R
\name{check_quantities}
\alias{check_quantities}
\alias{is_qty}
\alias{is_amount}
\alias{is_mass}
\alias{is_molecular_weight}
\alias{is_molarity_concentration}
\alias{is_mass_concentration}
\alias{is_volume}
\alias{is_pressure}
\alias{is_gas_solubility}
\alias{is_temperature}
\title{Check for quantity objects}
\usage{
is_qty(q)

is_amount(q)

is_mass(q)

is_molecular_weight(q)

is_molarity_concentration(q)

is_mass_concentration(q)

is_volume(q)

is_pressure(q)

is_gas_solubility(q)

is_temperature(q)
}
\arguments{
\item{q}{a quantity object}
}
\description{
These functions check whether something is a specific type of quantity.
}
\section{Functions}{
\itemize{
\item \code{is_qty}: check whether something is any quantity

\item \code{is_amount}: check whether something is an amount quantity

\item \code{is_mass}: check whether something is an amount quantity

\item \code{is_molecular_weight}: check whether something is a molecular weight quantity

\item \code{is_molarity_concentration}: check whether something is a molarity concentration quantity

\item \code{is_mass_concentration}: check whether something is a mass concentration quantity

\item \code{is_volume}: check whether something is a volume quantity

\item \code{is_pressure}: check whether something is a pressure quantity

\item \code{is_gas_solubility}: check whether something is a gas solubility quantity

\item \code{is_temperature}: check whether something is a temperature quantity
}}

\seealso{
Other quantity functions: 
\code{\link{get_quantity_info}},
\code{\link{metric_scaling}},
\code{\link{quantities}}
}
\concept{quantity functions}
