% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/chemtools.R
\docType{package}
\name{chemtools}
\alias{chemtools}
\alias{chemtools-package}
\title{chemtools package}
\description{
Collection of tools to simplify routine aquatic chemistry calculations.
}
\details{
This package includes a basic toolbox for units management with certain
types of \link{quantities} (\link{volume}, \link{concentration}, etc.) and full
support for \link{metric} scaling.
}
\author{
Sebastian Kopf
}
\seealso{
\link{quantity}, \link{metric}
}

