% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{constants}
\alias{constants}
\alias{ct_get_constant}
\alias{ct_constants}
\title{Constants}
\usage{
ct_get_constant(name)

ct_constants()
}
\arguments{
\item{name}{name of the constant}
}
\description{
List and retrieve constants used in chemtools.
}
\section{Functions}{
\itemize{
\item \code{ct_get_constant}: get the value of a constant

\item \code{ct_constants}: list all constants
}}

