% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities.R
\name{get_quantity_info}
\alias{get_quantity_info}
\alias{get_qty_value}
\alias{get_qty_text}
\alias{get_qty_units}
\alias{get_qty_units_with_label}
\title{Get quantity information}
\usage{
get_qty_value(q, unit = get_qty_units(q), transform = identity)

get_qty_text(q, unit = get_qty_units(q), signif = 5)

get_qty_units(q)

get_qty_units_with_label(q, label = names(q))
}
\arguments{
\item{q}{quantities}

\item{unit}{which units to retrieve quantity information in (by default the unit the quantity is in)}

\item{transform}{whether to transform the value with an additional function once in the desired units. Common transformation examples are log10 and log (natural log) but custom transformations are also possible. Default is NO transformation (\link{identity}).}

\item{signif}{number of significant digits for printing the quantity}

\item{label}{text label to use with the units - single value or vector of the same length as \code{q}. By default uses the names of \code{q}, which only works if \code{q} is a list or data frame.}
}
\description{
Get information about quantity objects including their values and text representations.
}
\section{Functions}{
\itemize{
\item \code{get_qty_value}: get the value of a quantity in the desired unit. By default returns the quantity in the units it is in.

\item \code{get_qty_text}: get the value of the quantity in the desired unit as a text string with the unit appended

\item \code{get_qty_units}: get units from a quantity, list of quantities or data frame (returns NA for objects/columns that are not quantities)

\item \code{get_qty_units_with_label}: get units from a quantity, list of quantities or data frame, with a custom label in the format \code{label [units]}. Objects/columns that are not quantities simply return the label with out the [units] part.
}}

\examples{
# quantity value examples
qty(0.1, "g") \%>\% get_qty_value()
qty(0.1, "g") \%>\% get_qty_value("g")
qty(0.1, "g") \%>\% get_qty_value("g", log10)
qty(0, "C") \%>\% get_qty_value("F")
qty(760, "Torr") \%>\% get_qty_value("atm")

# quantity text examples
qty(0.1, "g") \%>\% get_qty_text()
qty(0.1, "g") \%>\% get_qty_text("g")
qty(0:10, "C") \%>\% get_qty_text("F")
qty(760, "Torr") \%>\% get_qty_text("atm")

# quantity units examples
qty(5000, "g") \%>\% get_qty_units()
x <- list(a = qty(5000, "g"), b = 42, c = qty(100, "mbar"))
x \%>\% get_qty_units()
# labels with units
get_qty_units_with_label(qty(0.1, "mM"), "concentration")

# make labels with units for data frame columns
x <- data.frame(a = qty(1, "mg"), b = 2, c = qty(100, "mbar"))
get_qty_units_with_label(x)
get_qty_units_with_label(x, "same label")
}
\seealso{
Other quantity functions: 
\code{\link{check_quantities}},
\code{\link{metric_scaling}},
\code{\link{quantities}}
}
\concept{quantity functions}
