% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/aquasys.R
\name{is.aquasys}
\alias{is.aquasys}
\title{Reports whether x is any aquasys object (system, parameter or component)}
\usage{
is.aquasys(x)
}
\description{
Reports whether x is any aquasys object (system, parameter or component)
}

