% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{make_qty_text_labeller}
\alias{make_qty_text_labeller}
\title{Quantity labeller function}
\usage{
make_qty_text_labeller(q, unit = NULL, signif = 3)
}
\arguments{
\item{q}{the quantity for which to generate a text labeller}

\item{unit}{the unit (or units) in which to represent the quantity on the plot. By default uses the best metric prefix for each value of the quantity, which is good for generating orders of magnitude scales e.g. from mg to g to kg. If provided, uses the specified unit or units (need to be compatible with the quantity) for all values.}

\item{signif}{number of significant digits for printing the quantity}
}
\description{
This is a useful labeller function for ggplots to generate quantity text labels on discrete scales and facets in the used created by \code{\link{get_qty_text}}. Supports both best metric for each label (simply don't provide the \code{unit} parameter) or a specify unit (specify the \code{unit} parameter).
}
\examples{


# dual unit labels
q <- qty(seq(0, 1000, by = 100), "K")
l <- make_qty_text_labeller(q, unit = c("K", "C"))
l(q) # test how the labels would look
}
