% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{make_qty_units_explicit}
\alias{make_qty_units_explicit}
\title{Make units explicit}
\usage{
make_qty_units_explicit(df, ..., prefix = " [", suffix = "]")
}
\arguments{
\item{df}{the data frame in which to make the units explicit}

\item{...}{named arguments for the columns that should be made explicit and their units, e.g. \code{salt_weight = "mg", salt_concentration = "M"}.}

\item{prefix}{the prefix for the units}

\item{suffix}{the suffix for the units}
}
\description{
This function is intended for data frames / tibbles only and makes the specified columns explicit with the target units in the column name.
}
\examples{
# data frame with quantities
df <- tibble(x = qty(1:5, "mg"), y = qty(1:5, "L"))
df

# show with explicit units
make_qty_units_explicit(df, x = "mg")
make_qty_units_explicit(df, x = "g", y = "mL")
}
\seealso{
Other quantity functions: 
\code{\link{check_quantities}},
\code{\link{metric_scaling}},
\code{\link{quantities}},
\code{\link{quantity_data}},
\code{\link{quantity_units}}
}
\concept{quantity functions}
