% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{MediaChemTools-package}
\alias{MediaChemTools}
\alias{MediaChemTools-package}
\title{MediaChemTools: Culture media calculations.}
\description{
Package to simplify working with defined culture medium.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kopflab/mediachemtools}
  \item Report bugs at \url{https://github.com/kopflab/mediachemtools/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Kopf \email{sebastian.kopf@colorado.edu}

}
\keyword{internal}
