% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{metric}
\alias{metric}
\alias{cht_scale_metric}
\alias{cht_best_metric}
\alias{cht_base_metric}
\title{Metric prefixes}
\usage{
cht_scale_metric(q, prefix = "")

cht_best_metric(q)

cht_base_metric(q)
}
\arguments{
\item{q}{the \link{quantity} to scale}

\item{prefix}{a metric prefix (p, n, µ, m, k, M, etc.)}
}
\description{
These functions simplify converting between different metric prefixes.
}
\section{Functions}{
\itemize{
\item \code{cht_scale_metric}: scale to a specific metrix prefix (from whatever the quantity is currently in)

\item \code{cht_best_metric}: convert to best metric prefix (i.e. one that gives at least 1 significant digit before the decimal), 
if the quantity has a vector of values, scales to the best metric prefix for the median of all values

\item \code{cht_base_metric}: convert to base metric prefix of mediatools (i.e. to mol, L, etc.)
}}

\seealso{
Other functions: \code{\link{quantities}}
}
\concept{functions}
