% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{microbialkitchen-package}
\alias{microbialkitchen}
\alias{microbialkitchen-package}
\title{microbialkitchen: Culture Media Tools}
\description{
Package to simplify working with defined culture medium.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kopflab/microbialkitchen}
  \item Report bugs at \url{https://github.com/kopflab/microbialkitchen/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Kopf \email{sebastian.kopf@colorado.edu}

}
\keyword{internal}
