% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameter}
\alias{parameter}
\title{Set system parameters}
\usage{
parameter(...)
}
\arguments{
\item{...}{a list of parameter name and value pairings that modify
an existing aquatic system}
}
\description{
Use this function to modify system parameters
}
\details{
To see a graphical representation of the inheritance tree, see the
last example below.
}

