% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{quantities}
\alias{quantities}
\alias{qty}
\alias{get_qty_units}
\title{Quantities}
\usage{
qty(x, unit, scale_to_best_metric = TRUE)

get_qty_units(q)
}
\arguments{
\item{x}{the numeric value of the quantity, can be a single value or a vector}

\item{unit}{the unit of the quantity}

\item{scale_to_best_metric}{whether to automatically scale to the best metric prefix}

\item{q}{quantity or list of quantities}
}
\description{
The \code{qty} function makes it easy to keep track of different quantities in chemical calculations. Metric prefixes are fully supported, i.e. any unit can be combined with standard \link{metric} scaling (mL, nmol, µM, etc.). Some quantities can also be used in common \link{arithmetic} operations.
}
\details{
\emph{amount}: base unit \code{mol} but also understands \code{mole}, all metric prefixes allowed

\emph{mass}: base unit \code{g}, all metric prefixes allowed

\emph{molecular weight}: base unit \code{g/mol}, all metric prefixes allowed in the numerator

\emph{concentration (molarity)}: base unit \code{M} but also understands \code{mol/L}, all metric prefixes allowed in the numerator

\emph{concentration (density)}: base unit \code{g/L} but also understands \code{g/l}, all metric prefixes allowed in the numerator

\emph{volume}: base unit \code{L} but also understands \code{l}, all metric prefixes allowed

\emph{pressure}: base unit \code{bar} but also understands \code{Pa}, all metric prefixes allowed, the common non-metric units \code{atm} and \code{psi}, \code{Torr} and \code{mTorr} are also supported and will be automatically converted to \code{bar}

\emph{temperature}: base unit \code{K} but also understands \code{C} and \code{F} and converts them to Kelvin
}
\section{Functions}{
\itemize{
\item \code{qty}: generate a quantity object

\item \code{get_qty_units}: get units from a quantity or list of quantities (returns NA for objects that are not quantities)
}}

\examples{
qty(0.045, "mmol/L")
qty(6, "psi")
qty(30, "C")
}
\seealso{
Other functions: \code{\link{metric}}
}
\concept{functions}
