% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities.R
\name{quantity_data}
\alias{quantity_data}
\alias{get_qty_value}
\alias{as.double.microbial_kitchen_quantity}
\alias{get_qty_text}
\alias{as.character.microbial_kitchen_quantity}
\title{Get quantity data}
\usage{
get_qty_value(q, unit = get_qty_units(q), transform = identity)

\method{as.double}{microbial_kitchen_quantity}(x, unit = get_qty_units(x), ...)

get_qty_text(q, unit = get_qty_units(q), signif = 5)

\method{as.character}{microbial_kitchen_quantity}(x, unit = get_qty_units(x), ...)
}
\arguments{
\item{q}{quantities}

\item{unit}{which units to retrieve quantity data in (by default the unit the quantity is in)}

\item{transform}{whether to transform the value with an additional function once in the desired units. Common transformation examples are log10 and log (natural log) but custom transformations are also possible. Default is NO transformation (\link{identity}).}

\item{signif}{number of significant digits for printing the quantity}
}
\description{
Get data from quantity objects including their values (equivalent of \link[base]{as.numeric}) and text representations (equivalent of \link[base]{as.character}).
}
\section{Functions}{
\itemize{
\item \code{get_qty_value}: get the value of a quantity in the desired unit. By default returns the quantity in the units it is in. Functionally equivalent to \code{\link[base]{as.numeric}} and \code{\link[base]{as.double}}.

\item \code{as.double.microbial_kitchen_quantity}: S3 extension of \code{\link[base]{as.numeric}} and \link[base]{as.double} with optional \code{unit} argument

\item \code{get_qty_text}: get the value of the quantity in the desired unit as a text string with the unit appended. Functionally equivalent to \code{\link[base]{as.character}}.

\item \code{as.character.microbial_kitchen_quantity}: S3 implementation of \code{\link[base]{as.character}} with optional \code{unit} argument
}}

\examples{
# quantity value examples
qty(0.1, "g") \%>\% get_qty_value()
qty(0.1, "g") \%>\% get_qty_value("g")
qty(0.1, "g") \%>\% get_qty_value("g", log10)
qty(0, "C") \%>\% get_qty_value("F")
qty(760, "Torr") \%>\% get_qty_value("atm")
qty(760, "Torr") \%>\% as.numeric("atm")

# quantity text examples
qty(0.1, "g") \%>\% get_qty_text()
qty(0.1, "g") \%>\% get_qty_text("g")
qty(0:10, "C") \%>\% get_qty_text("F")
qty(760, "Torr") \%>\% get_qty_text("atm")
qty(760, "Torr") \%>\% as.character("atm")
}
\seealso{
Other quantity functions: 
\code{\link{check_quantities}},
\code{\link{make_qty_units_explicit}()},
\code{\link{metric_scaling}},
\code{\link{quantities}},
\code{\link{quantity_units}}
}
\concept{quantity functions}
