% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities.R
\name{quantity_info}
\alias{quantity_info}
\alias{get_qty_value}
\alias{get_qty_text}
\alias{get_qty_units}
\alias{get_qty_units_with_label}
\title{Get quantity information}
\usage{
get_qty_value(q, unit = get_qty_units(q))

get_qty_text(q, unit = get_qty_units(q), signif = 5)

get_qty_units(q)

get_qty_units_with_label(q, label = names(q))
}
\arguments{
\item{q}{quantity or list of quantities}

\item{unit}{the unit of the quantity}

\item{signif}{number of significant digits for printing the quantity}

\item{label}{text label to use with the units - single value or vector of the same length as \code{q}. By default uses the names of \code{q}, which only works if \code{q} is a list or data frame.}
}
\description{
Get quantity information
}
\details{
\code{get_qty_value}: get the value of a quantity in the desired unit. By default returns the quantity in the units it is in.

\code{get_qty_text}: get the value of the quantity in the desired unit as a text string with the unit appended

\code{get_qty_units}: get units from a quantity, list of quantities or data frame (returns NA for objects/columns that are not quantities)

\code{get_qty_units_with_label} get units from a quantity, list of quantities or data frame, with a custom label in the format \code{label [units]}. Objects/columns that are not quantities simply return the label with out the [units] part.
}
\examples{
# labels with units
get_qty_units_with_label(qty(0.1, "mM"), "concentration")

# make labels with units for data frame columns
x <- data.frame(a = qty(1, "mg"), b = 2, c = qty(100, "mbar"))
get_qty_units_with_label(x, names(x)) 
}
\seealso{
Other quantities: \code{\link{metric}},
  \code{\link{quantities}}
}
\concept{quantities}
