% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R, R/quantities.R, R/visualization.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tibble}
\alias{\%>\%}
\alias{as.factor}
\alias{as_factor}
\alias{scale_type}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{forcats}{\code{\link[forcats]{as_factor}}}

  \item{generics}{\code{\link[generics:coercion-factor]{as.factor}}}

  \item{ggplot2}{\code{\link[ggplot2]{scale_type}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{tibble}{\code{\link[tibble]{tibble}}}
}}

