% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbonate_chem.R
\name{speciation}
\alias{speciation}
\alias{calculate_DIC}
\alias{calculate_carbonic_acid}
\alias{calculate_bicarbonate}
\alias{calculate_carbonate}
\title{Speciation functions}
\usage{
calculate_DIC(pH, pCO2, `H2CO3*` = solubility * pCO2,
  solubility = calculate_solubility("CO2", temperature),
  temperature = qty(25, "C"), pKa1 = 6.3, pKa2 = 10.3)

calculate_carbonic_acid(pH, pCO2, DIC,
  solubility = calculate_solubility("CO2", temperature),
  temperature = qty(25, "C"), pKa1 = 6.3, pKa2 = 10.3)

calculate_bicarbonate(pH, pCO2, DIC,
  solubility = calculate_solubility("CO2", temperature),
  temperature = qty(25, "C"), pKa1 = 6.3, pKa2 = 10.3)

calculate_carbonate(pH, pCO2, DIC,
  solubility = calculate_solubility("CO2", temperature),
  temperature = qty(25, "C"), pKa1 = 6.3, pKa2 = 10.3)
}
\arguments{
\item{pH}{the pH of the system}

\item{pCO2}{the partial pressure of CO2 (a pressure quantity)}

\item{H2CO3*}{the concentration of H2CO3* (a molarity quantity), by default is calculated automatically from pCO2}

\item{solubility}{the solubility of CO2, by default is determined based on the temperature of the system}

\item{temperature}{the temperature of the system, used to calculate the solubility constant}

\item{pKa1}{the acid dissociation constant for H2CO3*}

\item{pKa2}{the acid dissociation constant for bicarbonate (HCO3-)}

\item{DIC}{dissolved inorganic carbon (a molarity quantity)}
}
\description{
Calculate dissolved inorganic carbon (DIC) and individual carbonate species in an open or closes system at equilibrium.
\itemize{
 \item{carbonic acid* = H2CO3* (aqueous CO2 + H2CO3)}
 \item{bicarbonate = HCO3(-)}
 \item{carbonate = CO3(2-)}  
}
}
\section{Functions}{
\itemize{
\item \code{calculate_DIC}: calculates the concentration of dissolved inorganic carbon based on pH and either pCO2 or H2CO3*. Returns DIC as a molarity quantity.

\item \code{calculate_carbonic_acid}: calculates the concentration of carbonic acid (H2CO3*) based on pH and either pCO2 or DIC. Returns carbonic acid as a molarity quantity.

\item \code{calculate_bicarbonate}: calculates the concentration of bicarbonate based on pH and either pCO2 or DIC. Returns bicarbonate concentration as a molarity quantity.

\item \code{calculate_carbonate}: calculates the concentration of carbonate based on pH and either pCO2 or DIC. Returns carbonate concentration as a molarity quantity.
}}

\examples{
calculate_DIC(pH = 7.0, pCO2 = qty(0.4, "mbar"))
calculate_DIC(pH = 7.0, pCO2 = qty(0.4, "mbar"), temperature = qty(0, "C"))
calculate_DIC(pH = 7.0, `H2CO3*` = qty(1, "mM"))
}
\seealso{
Other carbonate chemistry: \code{\link{closed_system}},
  \code{\link{open_system}}
}
\concept{carbonate chemistry}
