% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{vec_arith.mk_quantity}
\alias{vec_arith.mk_quantity}
\alias{vec_arith.mk_amount}
\alias{vec_arith.mk_mass}
\alias{vec_arith.mk_molecular_weight}
\alias{vec_arith.mk_molarity_concentration}
\alias{vec_arith.mk_mass_concentration}
\alias{vec_arith.mk_volume}
\alias{vec_arith.mk_pressure}
\alias{vec_arith.mk_gas_solubility}
\alias{vec_arith.mk_temperature}
\title{vec_arith for mk_quantity objects}
\usage{
\method{vec_arith}{mk_quantity}(op, x, y, ...)

\method{vec_arith}{mk_amount}(op, x, y, ...)

\method{vec_arith}{mk_mass}(op, x, y, ...)

\method{vec_arith}{mk_molecular_weight}(op, x, y, ...)

\method{vec_arith}{mk_molarity_concentration}(op, x, y, ...)

\method{vec_arith}{mk_mass_concentration}(op, x, y, ...)

\method{vec_arith}{mk_volume}(op, x, y, ...)

\method{vec_arith}{mk_pressure}(op, x, y, ...)

\method{vec_arith}{mk_gas_solubility}(op, x, y, ...)

\method{vec_arith}{mk_temperature}(op, x, y, ...)
}
\arguments{
\item{op}{An arithmetic operator as a string}

\item{x}{A pair of vectors. For \code{!}, unary \code{+} and unary \code{-}, \code{y} will be
a sentinel object of class \code{MISSING}, as created by \code{MISSING()}.}

\item{y}{A pair of vectors. For \code{!}, unary \code{+} and unary \code{-}, \code{y} will be
a sentinel object of class \code{MISSING}, as created by \code{MISSING()}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
vec_arith for mk_quantity objects

vec_arith for mk_amount objects

vec_arith for mk_mass objects

vec_arith for mk_molecular_weight objects

vec_arith for mk_molarity_concentration objects

vec_arith for mk_mass_concentration objects

vec_arith for mk_volume objects

vec_arith for mk_pressure objects

vec_arith for mk_gas_solubility objects

vec_arith for mk_temperature objects
}
