% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities.R
\name{vec_cast.microbial_kitchen_quantity}
\alias{vec_cast.microbial_kitchen_quantity}
\alias{vec_cast.microbial_kitchen_amount}
\alias{vec_cast.microbial_kitchen_mass}
\alias{vec_cast.microbial_kitchen_molecular_weight}
\alias{vec_cast.microbial_kitchen_molarity_concentration}
\alias{vec_cast.microbial_kitchen_mass_concentration}
\alias{vec_cast.microbial_kitchen_volume}
\alias{vec_cast.microbial_kitchen_pressure}
\alias{vec_cast.microbial_kitchen_gas_solubility}
\alias{vec_cast.microbial_kitchen_temperature}
\title{vec_cast for microbial_kitchen_quantity objects}
\usage{
\method{vec_cast}{microbial_kitchen_quantity}(x, to, ...)

\method{vec_cast}{microbial_kitchen_amount}(x, to, ...)

\method{vec_cast}{microbial_kitchen_mass}(x, to, ...)

\method{vec_cast}{microbial_kitchen_molecular_weight}(x, to, ...)

\method{vec_cast}{microbial_kitchen_molarity_concentration}(x, to, ...)

\method{vec_cast}{microbial_kitchen_mass_concentration}(x, to, ...)

\method{vec_cast}{microbial_kitchen_volume}(x, to, ...)

\method{vec_cast}{microbial_kitchen_pressure}(x, to, ...)

\method{vec_cast}{microbial_kitchen_gas_solubility}(x, to, ...)

\method{vec_cast}{microbial_kitchen_temperature}(x, to, ...)
}
\arguments{
\item{x}{Vectors to cast.}

\item{to}{Type to cast to. If \code{NULL}, \code{x} will be returned as is.}

\item{...}{For \code{vec_cast_common()}, vectors to cast. For
\code{vec_cast()}, \code{vec_cast_default()}, and \code{vec_restore()}, these
dots are only for future extensions and should be empty.}
}
\description{
vec_cast for microbial_kitchen_quantity objects

vec_cast for microbial_kitchen_amount objects

vec_cast for microbial_kitchen_mass objects

vec_cast for microbial_kitchen_molecular_weight objects

vec_cast for microbial_kitchen_molarity_concentration objects

vec_cast for microbial_kitchen_mass_concentration objects

vec_cast for microbial_kitchen_volume objects

vec_cast for microbial_kitchen_pressure objects

vec_cast for microbial_kitchen_gas_solubility objects

vec_cast for microbial_kitchen_temperature objects
}
