% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities.R
\name{vec_ptype2.microbial_kitchen_quantity}
\alias{vec_ptype2.microbial_kitchen_quantity}
\alias{vec_ptype2.microbial_kitchen_amount}
\alias{vec_ptype2.microbial_kitchen_mass}
\alias{vec_ptype2.microbial_kitchen_molecular_weight}
\alias{vec_ptype2.microbial_kitchen_molarity_concentration}
\alias{vec_ptype2.microbial_kitchen_mass_concentration}
\alias{vec_ptype2.microbial_kitchen_volume}
\alias{vec_ptype2.microbial_kitchen_pressure}
\alias{vec_ptype2.microbial_kitchen_gas_solubility}
\alias{vec_ptype2.microbial_kitchen_temperature}
\title{vec_ptype2 for microbial kitchen objects}
\usage{
\method{vec_ptype2}{microbial_kitchen_quantity}(x, y, ...)

\method{vec_ptype2}{microbial_kitchen_amount}(x, y, ...)

\method{vec_ptype2}{microbial_kitchen_mass}(x, y, ...)

\method{vec_ptype2}{microbial_kitchen_molecular_weight}(x, y, ...)

\method{vec_ptype2}{microbial_kitchen_molarity_concentration}(x, y, ...)

\method{vec_ptype2}{microbial_kitchen_mass_concentration}(x, y, ...)

\method{vec_ptype2}{microbial_kitchen_volume}(x, y, ...)

\method{vec_ptype2}{microbial_kitchen_pressure}(x, y, ...)

\method{vec_ptype2}{microbial_kitchen_gas_solubility}(x, y, ...)

\method{vec_ptype2}{microbial_kitchen_temperature}(x, y, ...)
}
\arguments{
\item{x}{Vector types.}

\item{y}{Vector types.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
vec_ptype2 for microbial kitchen objects
}
