% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{arithmetic}
\alias{arithmetic}
\title{Arithmetic}
\description{
These arithmetic operators for quantities in microbialkitchen keep track of units in standard operations (i.e. they ALL take the SI prefix into consideration). Also note that all operations that result in a new quantity object automatically scale the new value using \code{\link{best_metric}}.
}
