% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{arithmetic}
\alias{arithmetic}
\title{Arithmetic}
\usage{
qty == qty, qty != qty, qty < qty, qty <= qty, qty > qty, qty >= qty

qty + qty, qty - qty

qty / qty

amount / volume = molarity, amount / molarity = volume, molarity * volume = amount

mass / MW = amount, mass / amount = MW, amount * MW = mass
}
\description{
These arithmetic operators for quantities in chemtools keep track of units in standard operations (i.e. they ALL take the SI prefix into consideration). Also note that all operations that result in a new quantity object automatically scale the new value using \code{\link{best_metric}}.
}
\details{
\code{qty ==, !=, <, <=, >, >= qty} allows the comparison of quantities that are the same type (e.g. all mass).

\code{qty +- qty} allows the addition/subtraction of quantities that are the same type (e.g. all mass). Note that this operation also scales the new value using \code{\link{best_metric}}.

\code{qty / qty} divide quantities of the same type. Returns plain numeric (i.e. the units are divided out).

\code{amount / volume} divide an amount by a volume to create a molarity (concentration).

\code{amount / molarity} divide an amount by a molarity to create a volume.

\code{molarity * volume} multiply a molarity by a volume (or the other way around) to create an amount.

\code{mass / MW} divide a mass by a molecular weight to create an amount (mols).

\code{mass / amount} divice a mass by an amount (mols) to create a molecular weight.

\code{amount * MW} multiply an amount (mols) by a molecular weight to create a mass.
}
\examples{
qty(1, "g") == qty(1000, "mg") # TRUE
qty(2, "mg") < qty(5, "ng")  # FALSE
qty(1000, "mg") + qty(999, "g") # 1 kg
qty(5, "mg") / qty(1, "g") # 0.005
qty(5, "nmol") / qty(50, "mL") # 100 nM
qty(5, "nmol") / qty(100, "nM") # 50 mL
qty(100, "nM") * qty(50, "mL") # 5 nmol
qty(10, "g") / qty (50, "g/mol") # 200 mM 
qty(10, "g") / qty(200, "mmol") # 50 g/mol
qty(200, "mmol") * qty (50, "g/mol") # 10 g
}
