% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{constants}
\alias{constants}
\alias{get_microbialkitchen_constant}
\alias{get_microbialkitchen_constants}
\title{Constants}
\usage{
get_microbialkitchen_constant(name)

get_microbialkitchen_constants()
}
\arguments{
\item{name}{name of the constant}
}
\description{
List and retrieve constants used in microbialkitchen.
}
\section{Functions}{
\itemize{
\item \code{get_microbialkitchen_constant()}: get the value of a constant

\item \code{get_microbialkitchen_constants()}: list all constants

}}
