% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/constants.R
\name{constants}
\alias{constants}
\alias{get_constant}
\title{Constants}
\usage{
get_constant(name)

constants()
}
\arguments{
\item{name}{name of the constant}
}
\description{
List and retrieve constants used in chemtools.
}
\section{Functions}{
\itemize{
\item \code{get_constant}: get the value of a constant

\item \code{constants}: list all constants
}}

