% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{constants}
\alias{constants}
\alias{get_mediatools_constant}
\alias{get_mediatools_constants}
\title{Constants}
\usage{
get_mediatools_constant(name)

get_mediatools_constants()
}
\arguments{
\item{name}{name of the constant}
}
\description{
List and retrieve constants used in mediatools.
}
\section{Functions}{
\itemize{
\item \code{get_mediatools_constant}: get the value of a constant

\item \code{get_mediatools_constants}: list all constants
}}

