% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{constants}
\alias{constants}
\alias{cht_get_constant}
\alias{cht_constants}
\title{Constants}
\usage{
cht_get_constant(name)

cht_constants()
}
\arguments{
\item{name}{name of the constant}
}
\description{
List and retrieve constants used in mediatools.
}
\section{Functions}{
\itemize{
\item \code{cht_get_constant}: get the value of a constant

\item \code{cht_constants}: list all constants
}}

