% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mediatools-package}
\alias{mediatools}
\alias{mediatools-package}
\title{mediatools: Simplifies routine lab chem calculations.}
\description{
Helps with concentration calculations of gases and solutes in lab
    situations.
}
\keyword{internal}
