% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mediatools-package}
\alias{mediatools}
\alias{mediatools-package}
\title{mediatools: Simplify lab media calculations.}
\description{
Package to help with culturing media calculations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kopflab/mediatools}
  \item Report bugs at \url{https://github.com/kopflab/mediatools/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Kopf \email{sebastian.kopf@colorado.edu}

}
\keyword{internal}
