% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/units.R
\name{metric}
\alias{base_metric}
\alias{best_metric}
\alias{metric}
\alias{scale_metric}
\title{Metric prefixes}
\usage{
scale_metric(q, prefix = "")

best_metric(q)

base_metric(q)
}
\arguments{
\item{q}{the \link{quantity} to scale}

\item{prefix}{a metric prefix (p, n, µ, m, k, M, etc.)}
}
\description{
These functions simplify converting between different metric prefixes.
}
\section{Functions}{
\itemize{
\item \code{scale_metric}: scale to a specific metrix prefix (from whatever the quantity is currently in)

\item \code{best_metric}: convert to best metric prefix (i.e. one that gives at least 1 significant digit before the decimal),
if the quantity has a vector of values, scales to the best metric prefix for the median of all values

\item \code{base_metric}: convert to base metric prefix of chemtools (i.e. to mol, L, etc.)
}}
\seealso{
Other functions: \code{\link{amount}},
  \code{\link{concentration}}, \code{\link{mass}},
  \code{\link{molecular_weight}}, \code{\link{pressure}},
  \code{\link{qty}}, \code{\link{quantities}},
  \code{\link{quantity}}, \code{\link{temperature}},
  \code{\link{volume}}
}

