% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities.R
\name{quantities}
\alias{quantities}
\alias{qty}
\alias{is_qty}
\alias{is_amount}
\alias{is_mass}
\alias{is_molecular_mass}
\alias{is_molarity}
\alias{is_density}
\alias{is_volume}
\alias{is_pressure}
\alias{is_solubility}
\alias{is_temperature}
\title{Quantities}
\usage{
qty(x, unit, scale_to_best_metric = TRUE)

is_qty(q)

is_amount(q)

is_mass(q)

is_molecular_mass(q)

is_molarity(q)

is_density(q)

is_volume(q)

is_pressure(q)

is_solubility(q)

is_temperature(q)
}
\arguments{
\item{x}{the numeric value of the quantity, can be a single value or a vector}

\item{unit}{the unit of the quantity}

\item{scale_to_best_metric}{whether to automatically scale to the best metric prefix}

\item{q}{a quantity object}
}
\description{
The \code{qty} function makes it easy to keep track of different quantities in chemical calculations. Metric prefixes are fully supported, i.e. any unit can be combined with standard \link{metric} scaling (mL, nmol, µM, etc.). Some quantities can also be used in common \link{arithmetic} operations.
}
\details{
\emph{amount}: base unit \code{mol} but also understands \code{mole}, all metric prefixes allowed

\emph{mass}: base unit \code{g}, all metric prefixes allowed

\emph{molecular mass}: base unit \code{g/mol}, all metric prefixes allowed in the numerator

\emph{concentration (molarity)}: base unit \code{M} but also understands \code{mol/L}, all metric prefixes allowed in the numerator

\emph{concentration (density)}: base unit \code{g/L} but also understands \code{g/l}, all metric prefixes allowed in the numerator

\emph{volume}: base unit \code{L} but also understands \code{l}, all metric prefixes allowed

\emph{pressure}: base unit \code{bar} but also understands \code{Pa}, all metric prefixes allowed, the common non-metric units \code{atm} and \code{psi}, \code{Torr} and \code{mTorr} are also supported and will be automatically converted to \code{bar}

\emph{Henry's law solubility constant}: base unit \code{M/bar}, all metric prefixes allowed in the numerator, the common non-metric unit \code{M/atm} is also supported and will be automatically converted to \code{M/bar}.

\emph{temperature}: base unit \code{K} but also understands \code{C} and \code{F} and converts them to Kelvin
}
\section{Functions}{
\itemize{
\item \code{qty}: generate a quantity object

\item \code{is_qty}: check whether something is a quantity

\item \code{is_amount}: check whether something is an amount quantity

\item \code{is_mass}: check whether something is an amount quantity

\item \code{is_molecular_mass}: check whether something is a molecular mass quantity

\item \code{is_molarity}: check whether something is a molarity quantity

\item \code{is_density}: check whether something is a density quantity

\item \code{is_volume}: check whether something is a volume quantity

\item \code{is_pressure}: check whether something is a pressure quantity

\item \code{is_solubility}: check whether something is a solubility quantity

\item \code{is_temperature}: check whether something is a temperature quantity
}}

\examples{
qty(0.045, "mmol/L")
qty(6, "psi")
qty(30, "C")
}
\seealso{
Other quantities: \code{\link{metric}},
  \code{\link{quantity_info}}
}
\concept{quantities}
