% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities.R
\name{quantities}
\alias{quantities}
\alias{quantity}
\alias{qty}
\title{Quantities}
\usage{
qty(x = double(), unit, scale_to_best_metric = TRUE)
}
\arguments{
\item{x}{the numeric value of the quantity, can be a single value or a vector}

\item{unit}{the unit of the quantity}

\item{scale_to_best_metric}{whether to automatically scale to the best metric prefix}
}
\description{
The \code{qty} function makes it easy to keep track of different quantities in chemical calculations. Metric prefixes are fully supported, i.e. any unit can be combined with standard \link{metric} scaling (mL, nmol, kg, etc.). Some quantities can also be used in common \link{arithmetic} operations.
}
\details{
The following types of quantities are supported.

\emph{amount (N)}: base unit \code{mol} but also understands \code{mole}, all metric prefixes allowed

\emph{mass (m)}: base unit \code{g}, all metric prefixes allowed

\emph{molecular weight (MW)}: base unit \code{g/mol}, all metric prefixes allowed in the numerator

\emph{molarity concentration (C)}: base unit \code{M} but also understands \code{mol/L}, all metric prefixes allowed in the numerator

\emph{mass concentration (C)}: base unit \code{g/L} but also understands \code{g/l}, all metric prefixes allowed in the numerator. This is technically a density and sometimes referred to as such in the microbial kitchen documentation.

\emph{volume (V)}: base unit \code{L} but also understands \code{l}, all metric prefixes allowed

\emph{pressure (P)}: base unit \code{bar} but also understands \code{Pa}, all metric prefixes allowed, the common non-metric units \code{atm}, \code{psi}, \code{Torr}, \code{mTorr}, and \code{\% SP} (\% at standard pressure = \% of 1 bar) are also supported and will be automatically converted to \code{bar}.

\emph{gas solubility (S)}: base unit \code{M/bar}, all metric prefixes allowed in the numerator, the common non-metric unit \code{M/atm} is also supported and will be automatically converted to \code{M/bar}. This quantity is used for capturing Henry's law constants.

\emph{temperature (T)}: base unit \code{K} but also understands \code{C} and \code{F} and converts them to Kelvin
}
\section{Functions}{
\itemize{
\item \code{qty}: generate a quantity object
}}

\examples{
qty(0.045, "mmol/L")
qty(200, "mbar")
qty(6, "psi")
qty(30, "C")
qty(100, "K")
qty(5, "mg/L")
qty(1, "mM/bar")
qty(257, "g/mol")
}
\seealso{
Other quantity functions: 
\code{\link{check_quantities}},
\code{\link{get_quantity_info}},
\code{\link{metric_scaling}}
}
\concept{quantity functions}
