% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/units.R
\name{quantities}
\alias{amount}
\alias{concentration}
\alias{pressure}
\alias{qty}
\alias{quantities}
\alias{quantity}
\alias{temperature}
\alias{volume}
\title{Quantities}
\usage{
qty(x, unit, scale_to_best_metric = TRUE)

amount(x, unit, scale_to_best_metric = TRUE)

concentration(x, unit, scale_to_best_metric = TRUE)

volume(x, unit, scale_to_best_metric = TRUE)

pressure(x, unit, scale_to_best_metric = TRUE)

temperature(x, unit)
}
\arguments{
\item{x}{the numeric value of the quantity, can be a single value or a vector}

\item{unit}{the unit of the quantity}

\item{scale_to_best_metric}{whether to automatically scale to the best metric prefix}
}
\description{
These functions make it easy to keep track of different quantities in
chemical calculations. Metric prefixes are fully supported, i.e. any
unit can be combined with standard \link{metric} scaling (mL, nmol, µM, etc.)
}
\section{Functions}{
\itemize{
\item \code{qty}: universal function for generating a quantity (will try to guess which type from the unit)

\item \code{amount}: generate an amount (base unit \code{mol} but also understands \code{mole},
all metric prefixes allowed)

\item \code{concentration}: generate a molarity (base unit \code{M} but also understands \code{mol/L},
all metric prefixes allowed)

\item \code{volume}: generate a volume (base unit \code{L} but also understands \code{l},
all metric prefixes allowed)

\item \code{pressure}: generate a pressure (base unit \code{bar} but also understands \code{Pa},
all metric prefixes allowed, the common non-metric units \code{atm} and \code{psi}, \code{Torr} and
\code{mTorr} are also supported and will be automatically converted to \code{bar}).

\item \code{temperature}: generate a temperature (base unit \code{K} but also understands
\code{C} and \code{F} and converts them to Kelvin)
}}
\examples{
qty(0.045, "mmol/L")
qty(6, "psi")
qty(30, "C")
}
\seealso{
Other functions: \code{\link{base_metric}},
  \code{\link{best_metric}}, \code{\link{metric}},
  \code{\link{scale_metric}}
}

