% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities.R
\name{quantity_data}
\alias{quantity_data}
\alias{get_qty_value}
\alias{as.double.mk_quantity}
\alias{get_qty_text}
\alias{as.character.mk_quantity}
\alias{get_qty_text_each}
\title{Get quantity data}
\usage{
get_qty_value(x, unit = NULL, transform = identity)

\method{as.double}{mk_quantity}(x, unit = NULL, ...)

get_qty_text(x, unit = get_qty_units(x), signif = 5)

\method{as.character}{mk_quantity}(x, unit = get_qty_units(x), signif = 5, ...)

get_qty_text_each(x, signif = 5)
}
\arguments{
\item{x}{quantities}

\item{unit}{which units to retrieve quantity data in (by default the unit the quantity is in)}

\item{transform}{whether to transform the value with an additional function once in the desired units. Common transformation examples are log10 and log (natural log) but custom transformations are also possible. Default is NO transformation (\link{identity}).}

\item{signif}{number of significant digits for printing the quantity}
}
\description{
Get data from quantity objects including their values (equivalent of \link[base]{as.numeric}) and text representations (equivalent of \link[base]{as.character}).
}
\section{Functions}{
\itemize{
\item \code{get_qty_value}: get the value of a quantity in the desired unit. By default returns the quantity in the units it is in. Functionally equivalent to \code{\link[base]{as.numeric}} and \code{\link[base]{as.double}}.

\item \code{as.double.mk_quantity}: S3 extension of \code{\link[base]{as.numeric}} and \link[base]{as.double} with optional \code{unit} argument

\item \code{get_qty_text}: get the value of the quantity in the desired unit as a text string with the unit appended. Functionally equivalent to \code{\link[base]{as.character}}.

\item \code{as.character.mk_quantity}: S3 implementation of \code{\link[base]{as.character}} with optional \code{unit} and \code{signif} argument

\item \code{get_qty_text_each}: get each value of a quantity in the best metric unit with the unit appended. Note that if a value is zero, it will use the unit of the next smallest value for this number.
}}

\examples{
# quantity value examples
qty(0.1, "g") \%>\% get_qty_value()
qty(0.1, "g") \%>\% get_qty_value("g")
qty(0.1, "g") \%>\% get_qty_value("g", log10)
qty(0, "C") \%>\% get_qty_value("F")
qty(760, "Torr") \%>\% get_qty_value("atm")
qty(760, "Torr") \%>\% as.numeric("atm")

# quantity text examples
qty(0.1, "g") \%>\% get_qty_text()
qty(0.1, "g") \%>\% get_qty_text("g")
qty(0:10, "C") \%>\% get_qty_text("F")
qty(760, "Torr") \%>\% get_qty_text("atm")
qty(760, "Torr") \%>\% as.character("atm")

# each quantity text example
qty(c(0, 0.1, 10, 1000), "mg") \%>\% get_qty_text() # 0.1mg 10mg 1000mg
qty(c(0, 0.1, 10, 1000), "mg") \%>\% get_qty_text_each() # 0ug 100ug 10mg 1g
}
\seealso{
Other quantity functions: 
\code{\link{check_quantities}},
\code{\link{make_qty_units_explicit}()},
\code{\link{metric_scaling}},
\code{\link{quantities}},
\code{\link{quantity_units}}
}
\concept{quantity functions}
