% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{vec_arith.microbial_kitchen_quantity}
\alias{vec_arith.microbial_kitchen_quantity}
\alias{vec_arith.microbial_kitchen_amount}
\alias{vec_arith.microbial_kitchen_mass}
\alias{vec_arith.microbial_kitchen_molecular_weight}
\alias{vec_arith.microbial_kitchen_molarity_concentration}
\alias{vec_arith.microbial_kitchen_mass_concentration}
\alias{vec_arith.microbial_kitchen_volume}
\alias{vec_arith.microbial_kitchen_pressure}
\alias{vec_arith.microbial_kitchen_gas_solubility}
\title{vec_arith for microbial_kitchen_quantity objects}
\usage{
\method{vec_arith}{microbial_kitchen_quantity}(op, x, y, ...)

\method{vec_arith}{microbial_kitchen_amount}(op, x, y, ...)

\method{vec_arith}{microbial_kitchen_mass}(op, x, y, ...)

\method{vec_arith}{microbial_kitchen_molecular_weight}(op, x, y, ...)

\method{vec_arith}{microbial_kitchen_molarity_concentration}(op, x, y, ...)

\method{vec_arith}{microbial_kitchen_mass_concentration}(op, x, y, ...)

\method{vec_arith}{microbial_kitchen_volume}(op, x, y, ...)

\method{vec_arith}{microbial_kitchen_pressure}(op, x, y, ...)

\method{vec_arith}{microbial_kitchen_gas_solubility}(op, x, y, ...)
}
\arguments{
\item{op}{An arithmetic operator as a string}

\item{x}{A pair of vectors. For \code{!}, unary \code{+} and unary \code{-}, \code{y} will be
a sentinel object of class \code{MISSING}, as created by \code{MISSING()}.}

\item{y}{A pair of vectors. For \code{!}, unary \code{+} and unary \code{-}, \code{y} will be
a sentinel object of class \code{MISSING}, as created by \code{MISSING()}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
vec_arith for microbial_kitchen_quantity objects

vec_arith for microbial_kitchen_amount objects

vec_arith for microbial_kitchen_mass objects

vec_arith for microbial_kitchen_molecular_weight objects

vec_arith for microbial_kitchen_molarity_concentration objects

vec_arith for microbial_kitchen_mass_concentration objects

vec_arith for microbial_kitchen_volume objects

vec_arith for microbial_kitchen_pressure objects

vec_arith for microbial_kitchen_gas_solubility objects
}
