% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities.R
\name{vec_cast.mk_quantity}
\alias{vec_cast.mk_quantity}
\alias{vec_cast.mk_amount}
\alias{vec_cast.mk_mass}
\alias{vec_cast.mk_molecular_weight}
\alias{vec_cast.mk_molarity_concentration}
\alias{vec_cast.mk_mass_concentration}
\alias{vec_cast.mk_volume}
\alias{vec_cast.mk_pressure}
\alias{vec_cast.mk_gas_solubility}
\alias{vec_cast.mk_temperature}
\title{vec_cast for mk_quantity objects}
\usage{
\method{vec_cast}{mk_quantity}(x, to, ...)

\method{vec_cast}{mk_amount}(x, to, ...)

\method{vec_cast}{mk_mass}(x, to, ...)

\method{vec_cast}{mk_molecular_weight}(x, to, ...)

\method{vec_cast}{mk_molarity_concentration}(x, to, ...)

\method{vec_cast}{mk_mass_concentration}(x, to, ...)

\method{vec_cast}{mk_volume}(x, to, ...)

\method{vec_cast}{mk_pressure}(x, to, ...)

\method{vec_cast}{mk_gas_solubility}(x, to, ...)

\method{vec_cast}{mk_temperature}(x, to, ...)
}
\arguments{
\item{x}{Vectors to cast.}

\item{to}{Type to cast to. If \code{NULL}, \code{x} will be returned as is.}

\item{...}{For \code{vec_cast_common()}, vectors to cast. For
\code{vec_cast()}, \code{vec_cast_default()}, and \code{vec_restore()}, these
dots are only for future extensions and should be empty.}
}
\description{
vec_cast for mk_quantity objects

vec_cast for mk_amount objects

vec_cast for mk_mass objects

vec_cast for mk_molecular_weight objects

vec_cast for mk_molarity_concentration objects

vec_cast for mk_mass_concentration objects

vec_cast for mk_volume objects

vec_cast for mk_pressure objects

vec_cast for mk_gas_solubility objects

vec_cast for mk_temperature objects
}
